minor re-org, add metrics to config reloader
