--- config/file.go
@@ -1,41 +1,0 @@
-// Copyright 2023 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package config
-
-import (
-	"fmt"
-	"os"
-
-	"gopkg.in/yaml.v2"
-)
-
-func FromFile(path string) (*Config, error) {
-	configFromFile := &Config{}
-
-	var bs []byte
-	var err error
-	if bs, err = os.ReadFile(path); err != nil {
-		return nil, fmt.Errorf("failed to load %s: %w", path, err)
-	}
-
-	if err = yaml.Unmarshal(bs, configFromFile); err != nil {
-		return nil, fmt.Errorf("failed to parse %s: %w", path, err)
-	}
-
-	if err = configFromFile.Validate(); err != nil {
-		return nil, fmt.Errorf("config is invalid %s: %w", path, err)
-	}
-
-	return configFromFile, err
-}
